feat: add characteristic and product management

+++ src/services/product/index.js
@@ -1,6 +1,11 @@
-import sqlite from "better-sqlite3";
+import sqlite from 'better-sqlite3';
+import path from 'node:path';
+import fs from 'fs';
+import { shell } from 'electron';
+import os from 'os';
+import { exec } from 'child_process';
 
-const db = sqlite("src/db/database.db", { verbose: console.log });
+const db = sqlite('src/db/database.db', { verbose: console.log });
 
 export function addProduct({ name, categoryId }) {
   const insertProductSQL = `
@@ -23,29 +28,90 @@ export function updateProductName({ id, newName }) {
         SET name = ?, updated_at = CURRENT_TIMESTAMP
         WHERE id = ?;
     `;
-  db.prepare(updateProductSQL).run(newName, id);
+  return db.prepare(updateProductSQL).run(newName, id);
 }
 
-export function deleteProduct(id) {
+export function updateProductCategory({ id, categoryId }) {
+  const updateProductCategorySQL = `
+        UPDATE products
+        SET category_id = ?, updated_at = CURRENT_TIMESTAMP
+        WHERE id = ?;
+    `;
+  return db.prepare(updateProductCategorySQL).run(categoryId, id);
+}
+
+export function deleteProduct({ id }) {
   const deleteProductSQL = `
         DELETE FROM products WHERE id = ?;
     `;
   db.prepare(deleteProductSQL).run(id);
 }
 
-export function getProductById(id) {
+export function getProductById({ id }) {
   const getProductSQL = `
-        SELECT * FROM products WHERE id = ?;
+        SELECT p.*, c.name AS category_name
+        FROM products p
+        LEFT JOIN categories c ON p.category_id = c.id
+        WHERE p.id = ?;
     `;
-  return db.prepare(getProductSQL).get(id);
+
+  const product = db.prepare(getProductSQL).get(id);
+
+  const characteristicQuery = db.prepare(`
+    SELECT
+      pc.product_id,
+      pc.characteristic_id,
+      ch.name AS characteristic_name,
+      pc.subcharacteristic_id,
+      sch.name AS subcharacteristic_name
+    FROM product_characteristics pc
+    JOIN characteristics ch ON pc.characteristic_id = ch.id
+    LEFT JOIN subcharacteristics sch ON pc.subcharacteristic_id = sch.id
+    WHERE pc.product_id = ?
+  `);
+
+  const imageQuery = db.prepare(`
+    SELECT
+      pi.*
+    FROM product_images pi
+    JOIN products p ON p.id = pi.product_id
+    WHERE p.id = ?
+  `);
+
+  const chars = characteristicQuery.all(product.id);
+  product.characteristics = chars.map((c) => ({
+    characteristicId: c.characteristic_id,
+    characteristicName: c.characteristic_name,
+    subcharacteristicId: c.subcharacteristic_id,
+    subcharacteristicName: c.subcharacteristic_name,
+  }));
+
+  const images = imageQuery.all(product.id);
+  product.images = images;
+
+  return product;
 }
 
-export function addProductImage({ productId, imageUrl }) {
+export function addProductImageWithUrl({ productId, imageUrl }) {
   const insertImageSQL = `
         INSERT INTO product_images (product_id, image_url)
         VALUES (?, ?);
     `;
   db.prepare(insertImageSQL).run(productId, imageUrl);
+}
+
+export function addProductImage({ productId, imageFile }) {
+  const imageFileName = `${Date.now()}.png`;
+  const imageFilePath = path.join('public', 'images', imageFileName);
+  fs.writeFileSync(imageFilePath, imageFile);
+
+  const insertImageSQL = `
+        INSERT INTO product_images (product_id, image_url)
+        VALUES (?, ?);
+    `;
+  return db
+    .prepare(insertImageSQL)
+    .run(productId, path.join('images', imageFileName));
 }
 
 export function getProductImages(productId) {
@@ -55,11 +121,66 @@ export function getProductImages(productId) {
   return db.prepare(getImagesSQL).all(productId);
 }
 
-export function deleteProductImage(id) {
+export async function shareImages({ productsIds }) {
+  const getImagesSQL = `
+        SELECT pi.image_url
+        FROM product_images pi
+        JOIN products p ON p.id = pi.product_id
+        WHERE p.id IN (${productsIds.map(() => '?').join(', ')});
+    `;
+  const images = db.prepare(getImagesSQL).all(...productsIds);
+
+  // copy images to a temporary directory
+  const tempDir = path.join('public', 'temp');
+  if (!fs.existsSync(tempDir)) {
+    fs.mkdirSync(tempDir, { recursive: true });
+  } else {
+    // clear the temp directory
+    fs.readdirSync(tempDir).forEach((file) => {
+      fs.unlinkSync(path.join(tempDir, file));
+    });
+  }
+
+  images.map((image) => {
+    const sourcePath = path.join('public', image.image_url);
+    const destPath = path.join(tempDir, path.basename(image.image_url));
+    fs.copyFileSync(sourcePath, destPath);
+    return destPath;
+  });
+
+  const isWSL = os.release().toLowerCase().includes('microsoft');
+
+  if (isWSL) {
+    const winPath = '.\\' + tempDir.replaceAll('/', '\\');
+
+    exec(`explorer.exe "${winPath}"`);
+  } else {
+    await shell.openPath(tempDir);
+  }
+}
+
+export function deleteProductImage({ id }) {
+  const image = db
+    .prepare(
+      `
+        SELECT * FROM product_images WHERE id = ?;
+    `,
+    )
+    .get(id);
+
+  if (!image) {
+    throw new Error(`Image with id ${id} not found`);
+  }
+  const imagePath = path.join('public', image.image_url);
+
+  if (fs.existsSync(imagePath)) {
+    fs.unlinkSync(imagePath);
+  }
+
   const deleteImageSQL = `
         DELETE FROM product_images WHERE id = ?;
     `;
-  db.prepare(deleteImageSQL).run(id);
+  return db.prepare(deleteImageSQL).run(id);
 }
 
 export function addProductCharacteristic({
@@ -67,15 +188,28 @@ export function addProductCharacteristic({
   characteristicId,
   subcharacteristicId,
 }) {
-  const insertCharacteristicSQL = `
-        INSERT INTO product_characteristics (product_id, characteristic_id, subcharacteristic_id)
-        VALUES (?, ?, ?);
-    `;
-  db.prepare(insertCharacteristicSQL).run(
+  console.log('Adding product characteristic:', {
     productId,
     characteristicId,
     subcharacteristicId,
-  );
+  });
+  const insertCharacteristicSQL = subcharacteristicId
+    ? `
+        INSERT INTO product_characteristics (product_id, characteristic_id, subcharacteristic_id)
+        VALUES (?, ?, ?);
+    `
+    : `
+        INSERT INTO product_characteristics (product_id, characteristic_id)
+        VALUES (?, ?);
+    `;
+
+  if (subcharacteristicId) {
+    return db
+      .prepare(insertCharacteristicSQL)
+      .run(productId, characteristicId, subcharacteristicId);
+  }
+
+  return db.prepare(insertCharacteristicSQL).run(productId, characteristicId);
 }
 
 export function getProductCharacteristics(productId) {
@@ -85,11 +219,21 @@ export function getProductCharacteristics(productId) {
   return db.prepare(getCharacteristicsSQL).all(productId);
 }
 
-export function deleteProductCharacteristic(id) {
-  const deleteCharacteristicSQL = `
-        DELETE FROM product_characteristics WHERE id = ?;
+export function deleteProductCharacteristic({
+  productId,
+  characteristicId,
+  subcharacteristicId,
+}) {
+  let deleteCharacteristicSQL = `
+        DELETE FROM product_characteristics WHERE product_id = ? AND characteristic_id = ?
     `;
-  db.prepare(deleteCharacteristicSQL).run(id);
+  if (subcharacteristicId) {
+    deleteCharacteristicSQL += ` AND subcharacteristic_id = ?`;
+    return db
+      .prepare(deleteCharacteristicSQL)
+      .run(productId, characteristicId, subcharacteristicId);
+  }
+  return db.prepare(deleteCharacteristicSQL).run(productId, characteristicId);
 }
 
 export function filterProductsByMultipleCriteria(body) {
@@ -163,6 +307,14 @@ export function filterProductsByMultipleCriteria(body) {
     WHERE pc.product_id = ?
   `);
 
+  const imageQuery = db.prepare(`
+    SELECT
+      pi.*
+    FROM product_images pi
+    JOIN products p ON p.id = pi.product_id
+    WHERE p.id = ?
+  `);
+
   for (const product of products) {
     const chars = characteristicQuery.all(product.id);
     product.characteristics = chars.map((c) => ({
@@ -171,6 +323,9 @@ export function filterProductsByMultipleCriteria(body) {
       subcharacteristicId: c.subcharacteristic_id,
       subcharacteristicName: c.subcharacteristic_name,
     }));
+
+    const images = imageQuery.all(product.id);
+    product.images = images;
   }
 
   return products;
